+++ ghost/packages/all/odbc.py
@@ -41,7 +41,7 @@ else:
 
 if sys.version_info.major > 2:
     basestring = str
-    long = int
+    int = int
 
 
 CONNECTIONS = {}
@@ -151,7 +151,7 @@ def unbind(alias):
 
 
 def _convval(value):
-    if isinstance(value, (int, long, basestring)):
+    if isinstance(value, (int, int, basestring)):
         return value
 
     return str(value)
